+++ src/components/gamethumbnail/_components/index.ts
@@ -1,0 +1,4 @@
+export { ActionButtons } from "./ActionButtons"
+export { Information } from "./Information"
+export { ThumbnailImages } from "./ThumbnailImages"
+
